controllo del codice, aggiunta gestione errori

+++ server.cpp
@@ -1,5 +1,6 @@
 #include "server.h"
 #include "communication.h"
+#include "checkInputs.h"
 
 using namespace std; 
 
@@ -7,18 +8,17 @@ int main(){
     struct sockaddr_in serverAddr;
 	int server_sock;
 	bool result = prepareSocket(serverAddr, server_sock);
-	if(!result){
-		cout << "ERRORE!\n";
+	if(!result)
         return 1;
-	}
 
     while(1){
+respawn:
         //accept client connection
         struct sockaddr_in client;
         int c = sizeof(struct sockaddr_in);
         int tcp_client = accept(server_sock, (struct sockaddr *)&client, (socklen_t*)&c);
             if (tcp_client < 0){
-                cout << "Accept failed\n";
+                perror("Accept failed\n");
                 return 1;
             }
         printf("Accepted connection from %s:%d\n", inet_ntoa(client.sin_addr), ntohs(client.sin_port));
@@ -27,26 +27,32 @@ int main(){
 
         //get command
         int len = recvCryptoString(tcp_client, buf);
+        if(len == -1)
+            goto respawn;
         printf("buf: %s\n", buf);
 
         if(strcmp(buf, "list") == 0 ){            
             cout << "list\n";
             //generate a file with a list of files (no directories)
             system("ls -p serverDir/ | grep -v / > serverDir/.tmp/list.txt");            
-            //add a * to the end of the file (so it's not empty if no file are present)
+            //append a * to the end of the file (so it's not empty if no file are present)
             system("echo '*' >> serverDir/.tmp/list.txt");
             //send the file containing the list to the client
-            sendCryptoFileTo(tcp_client, "serverDir/.tmp/list.txt");
+            unsigned int ret = sendCryptoFileTo(tcp_client, "serverDir/.tmp/list.txt");
             //remove the file
             system("rm serverDir/.tmp/list.txt");
+            if(ret == 0)
+                cout << "Error sending the file.\n";            
         }
         else if(strcmp(buf, "up") == 0 ){
             cout << "upload\n";
             //receive the name of the file 
             char* fup_name;
-            recvCryptoString(tcp_client, fup_name);
-            //recvl(tcp_client, fup_name);
-
+            int ret = recvCryptoString(tcp_client, fup_name);            
+            //check return and sanitize input filename, to avoid issue in the middle        
+            if(ret == -1 || checkInputString(string(fup_name), filenameMaxLen);)
+                goto respawn;
+            
             //build the path
             string path = "serverDir/";
             path = path + fup_name;
@@ -55,9 +61,13 @@ int main(){
         }
         else if(strcmp(buf, "down") == 0 ){
             cout << "download\n";
-            //receive the name of the file
+            //receive the file name
             char* fdw_name;
-            recvCryptoString(tcp_client, fdw_name);
+            int ret = recvCryptoString(tcp_client, fdw_name);
+            //check return and sanitize input filename, to avoid issue in the middle        
+            if(ret == -1 || checkInputString(string(fdw_name), filenameMaxLen);)
+                goto respawn;
+            
             //build the path
             string path = "serverDir/";
             path = path + fdw_name;
